renaming message.rs to api_client.rs so things look cleaner.
